Fix Jump To Top Button Issue
Fix an issue in safari mobile and internet explorer that prevents the
jump to top button from staying fixed at the bottom right of the
viewport. The issue causes the button to remain fixed to its original
location. The cause of the issue is simply that the browsers this
issue affects do not support sticky positioning.